Add hooks before saving resource
Partly addresses
https://github.com/apokalipto/devise_saml_authenticatable/issues/38

This PR introduces a mechanism to apply custom behaviour prior to saving
a resource. This behaviour occurs when `Devise.saml_update_user` or
`Devise.saml_create_user` has been enabled, and the resource has been
found or created. The mechanism works by means of introducing a hook
just prior to saving the resource, providing an application the
flexibility to make arbitrary changes to the resource.

This hook replaces the existing `set_user_saml_attributes` with a
default hook implementation which can be replaced or extended depending
on the application needs.

The changes are backward compatible to my knowledge, public
configuration has not been modified, but a new configuration has been
introduced (`saml_update_resource_hook`).