Ability to set attribute_map via config setting
Instead of storing the attribute_map in attribute-map.yml, store it in the database, or set it programmatically ...

Thank you for your lovely gem. It is very useful to me.

In our case, our customers will be configuring SAML via a configuration panel. They will be setting up the attribute map themselves. If this is stored in the database, we can use classic Rails methods to create a nicer UI for them. This way they don't need to go edit some strange file in the guts of our product.

Thanks again!

Fixes #65